Rename warc.{c,h} to htswarc.{c,h} for the hts* file convention
First-party engine sources are hts-prefixed; only vendored code (md5,
punycode, minizip, coucal) is not. Rename to match, and update the includers
and both build files (Makefile.am, libhttrack.vcxproj).
