+++ src/htswarc.h
